Node JS vewrsion upgrade
AWS started sending emails warning us that nodeJS 18x was being deprecated.  This change updates the version of nodeJS that we target to 22x.
Also fixes up a bunmch of aws-CLI and CDK versions to work again after a dev container feature (ghcr.io/devcontainers-contrib/features/aws-cdk) that we relied on disappeared